libnm-core: return NULL for unset boxed properties in wired setting
Boxed properties have a default value of NULL. If the getter returns
anything different from NULL when the value is not set, for example an
empty GHashTable or strv, then a nm_setting_diff() between a newly
created setting and NULL with
NM_SETTING_COMPARE_FLAG_DIFF_RESULT_NO_DEFAULT will report the
property as changed even if nothing really changed.

Update the boxed properties in the wired setting to return NULl when
they are not set.