libnm-core,core: add and use now setting diff flag in reapply
When reapplying a connection we do a diff between the old and new
connection. Since we don't pass the DIFF_RESULT_NO_DEFAULT comparison
flag, all the properties of a setting that is present in only one
connection are reported as changed.

This is not always correct: for some settings (for example, the wired
setting for a bridge connection) the mere presence of the setting
doesn't make a difference if all its properties are set to
default. For other settings like SR-IOV there is a difference (the
setting with all default values means "manage the SR-IOV parameters
and set them to default values", while no setting means "don't manage
them").

Add a new comparison flag DIFF_EFFECTIVE that uses knowledge from each
setting to tell what changed in the connection taking into account the
considerations above.